more Java tests for System Key
