+++ AwsCryptographicMaterialProviders/runtimes/java/src/testExamples/java/software/amazon/cryptography/example/hierarchy/mutations/MutationKmsAccessTerminalInFlightTest.java
@@ -1,6 +1,6 @@
 // Copyright Amazon.com Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-package software.amazon.cryptography.example.hierarchy;
+package software.amazon.cryptography.example.hierarchy.mutations;
 
 import static software.amazon.cryptography.example.Fixtures.MRK_ARN_WEST;
 import static software.amazon.cryptography.example.Fixtures.POSTAL_HORN_KEY_ARN;
@@ -19,7 +19,9 @@ import software.amazon.awssdk.services.kms.model.KmsException;
 import software.amazon.cryptography.example.CredentialUtils;
 import software.amazon.cryptography.example.DdbHelper;
 import software.amazon.cryptography.example.Fixtures;
-import software.amazon.cryptography.example.hierarchy.mutations.MutationsProvider;
+import software.amazon.cryptography.example.StorageCheater;
+import software.amazon.cryptography.example.hierarchy.AdminProvider;
+import software.amazon.cryptography.example.hierarchy.CreateKeyExample;
 import software.amazon.cryptography.keystore.KeyStorageInterface;
 import software.amazon.cryptography.keystoreadmin.KeyStoreAdmin;
 import software.amazon.cryptography.keystoreadmin.model.ApplyMutationInput;
@@ -36,42 +38,36 @@ import software.amazon.cryptography.keystoreadmin.model.Mutations;
 import software.amazon.cryptography.keystoreadmin.model.SystemKey;
 import software.amazon.cryptography.keystoreadmin.model.TrustStorage;
 
-public class MutationKmsAccessOriginalInFlightTest {
+public class MutationKmsAccessTerminalInFlightTest {
 
   static final String testPrefix =
-    "mutation-kms-access-in-flight-original-test-";
+    "mutation-kms-access-in-flight-terminal-test-";
 
   static final Pattern matchBranchKeyType = Pattern.compile(
     "(?<=Branch Key Type: )(.*)(?:;)"
   );
 
   @Test
-  public void test() throws InterruptedException {
-    SystemKey systemKey = SystemKey
-      .builder()
-      .trustStorage(TrustStorage.builder().build())
-      .build();
-    KeyStorageInterface storage = StorageExample.create(
+  public void test() {
+    KeyStorageInterface storage = StorageCheater.create(
       Fixtures.ddbClientWest2,
       Fixtures.TEST_KEYSTORE_NAME,
       Fixtures.TEST_LOGICAL_KEYSTORE_NAME
     );
-    // KeyStorageInterface storage = StorageCheater.create(
-    //   Fixtures.ddbClientWest2,
-    //   Fixtures.TEST_KEYSTORE_NAME,
-    //   Fixtures.TEST_LOGICAL_KEYSTORE_NAME
-    // );
+    SystemKey systemKey = SystemKey
+      .builder()
+      .trustStorage(TrustStorage.builder().build())
+      .build();
     final String branchKeyId =
       testPrefix + java.util.UUID.randomUUID().toString();
-
     CreateKeyExample.CreateKey(
       Fixtures.TEST_KEYSTORE_NAME,
       Fixtures.TEST_LOGICAL_KEYSTORE_NAME,
-      MRK_ARN_WEST,
+      POSTAL_HORN_KEY_ARN,
       branchKeyId,
       Fixtures.ddbClientWest2
     );
-    KeyManagementStrategy strategyAll = AdminProvider.strategy(
+    KeyManagementStrategy strategyWest2 = AdminProvider.strategy(
       Fixtures.kmsClientWest2
     );
     KmsClient denyMrk = KmsClient
@@ -91,25 +87,26 @@ public class MutationKmsAccessOriginalInFlightTest {
 
     KeyManagementStrategy strategyDenyMrk = AdminProvider.strategy(denyMrk);
     KeyStoreAdmin admin = AdminProvider.admin(
+      Fixtures.TEST_KEYSTORE_NAME,
       Fixtures.TEST_LOGICAL_KEYSTORE_NAME,
-      storage
+      Fixtures.ddbClientWest2
     );
 
     System.out.println("BranchKey ID to mutate: " + branchKeyId);
-    HashMap<String, String> terminalEC = new HashMap<>();
+    HashMap<String, String> terminalEC = new HashMap<>(2, 1);
     terminalEC.put("Robbie", "is a dog.");
 
     Mutations mutations = Mutations
       .builder()
       .TerminalEncryptionContext(terminalEC)
-      .TerminalKmsArn(POSTAL_HORN_KEY_ARN)
+      .TerminalKmsArn(MRK_ARN_WEST)
       .build();
 
     InitializeMutationInput initInput = InitializeMutationInput
       .builder()
       .Mutations(mutations)
       .Identifier(branchKeyId)
-      .Strategy(strategyAll)
+      .Strategy(strategyWest2)
       .SystemKey(systemKey)
       .build();
 
@@ -121,19 +118,14 @@ public class MutationKmsAccessOriginalInFlightTest {
       " items: \n" +
       MutationsProvider.mutatedItemsToString(initOutput.MutatedBranchKeyItems())
     );
+
     boolean done = false;
     List<Exception> exceptions = new ArrayList<>();
     boolean isFromThrown = false;
     boolean isToThrown = false;
-    boolean verifyTerminalThrown = false;
     int limitLoop = 5;
 
     while (!done) {
-      // System.out.println(
-      //   "Loop Count': " +
-      //   limitLoop +
-      //   " Sleeping 10 seconds\n");
-      // Thread.sleep(10000);
       try {
         limitLoop--;
         if (limitLoop == 0) done = true;
@@ -147,7 +139,7 @@ public class MutationKmsAccessOriginalInFlightTest {
         ApplyMutationOutput applyOutput = admin.ApplyMutation(applyInput);
         ApplyMutationResult result = applyOutput.MutationResult();
         System.out.println(
-          "\nApplyLogs: " +
+          "ApplyLogs: " +
           branchKeyId +
           " items: \n" +
           MutationsProvider.mutatedItemsToString(
@@ -186,6 +178,7 @@ public class MutationKmsAccessOriginalInFlightTest {
             accessDenied
           );
         }
+        // An exception was thrown, let's delete the item
         if (accessDenied.getMessage().contains("branch:version")) {
           Matcher matcher = matchBranchKeyType.matcher(
             accessDenied.getMessage()
@@ -219,8 +212,8 @@ public class MutationKmsAccessOriginalInFlightTest {
     // Clean Up
     Fixtures.cleanUpBranchKeyId(storage, branchKeyId);
     Assert.assertTrue(
-      (exceptions.size() == 1),
-      "Only 1 exceptions should have been thrown. But got " +
+      (exceptions.size() == 2),
+      "Only two exceptions should have been thrown. But got " +
       exceptions.size() +
       ". Exceptions:\n" +
       exceptions
@@ -228,10 +221,10 @@ public class MutationKmsAccessOriginalInFlightTest {
         .map(Throwable::toString)
         .collect(Collectors.joining("\n"))
     );
+    Assert.assertTrue(isToThrown, "MutationToException MUST be thrown.");
     Assert.assertFalse(
-      isToThrown,
-      "MutationToException should never be thrown."
+      isFromThrown,
+      "MutationFromException should never be thrown."
     );
-    Assert.assertTrue(isFromThrown, "MutationFromException MUST be thrown.");
   }
 }
