+++ AwsCryptographicMaterialProviders/runtimes/java/src/testExamples/java/software/amazon/cryptography/example/hierarchy/mutations/ScanForInFlightMutationsTest.java
@@ -1,8 +1,7 @@
-package software.amazon.cryptography.example.hierarchy;
+package software.amazon.cryptography.example.hierarchy.mutations;
 
 import org.testng.annotations.Test;
 import software.amazon.cryptography.example.Fixtures;
-import software.amazon.cryptography.example.hierarchy.mutations.ScanForInFlightMutations;
 import software.amazon.cryptography.example.hierarchy.mutations.ScanForInFlightMutations.PageResult;
 
 public class ScanForInFlightMutationsTest {
